--- src/@lib/hooks/useAppContext.ts
@@ -1,12 +1,0 @@
-import { createContext, useContext } from "react";
-import { AppContextType } from "../../types";
-
-export const AppContext = createContext<AppContextType | undefined>(undefined);
-
-export const useAppContext = () => {
-  const context = useContext(AppContext);
-  if (context === undefined) {
-    throw new Error("useAppContext must be used within an AppProvider");
-  }
-  return context;
-};
