+++ src/components/Header.tsx
@@ -1,9 +1,10 @@
-import { useAppContext } from "../@lib/hooks/useAppContext";
+import { useAppContext, useThemeContext } from "../@lib/hooks/useContext";
 import { renderLog } from "../utils";
 
 export const Header: React.FC = () => {
   renderLog("Header rendered");
-  const { theme, toggleTheme, user, login, logout } = useAppContext();
+  const { theme, toggleTheme } = useThemeContext();
+  const { user, login, logout } = useAppContext();
 
   const handleLogin = () => {
     // 실제 애플리케이션에서는 사용자 입력을 받아야 합니다.
