+++ src/components/ItemList.tsx
@@ -1,15 +1,17 @@
 import { useState } from "react";
-import { useAppContext } from "../@lib/hooks/useAppContext";
+
 import { Item } from "../types";
 import { renderLog } from "../utils";
+import { useThemeContext } from "../@lib/hooks/useContext";
+import { memo } from "../@lib";
 
 export const ItemList: React.FC<{
   items: Item[];
   onAddItemsClick: () => void;
-}> = ({ items, onAddItemsClick }) => {
+}> = memo(({ items, onAddItemsClick }) => {
   renderLog("ItemList rendered");
   const [filter, setFilter] = useState("");
-  const { theme } = useAppContext();
+  const { theme } = useThemeContext();
 
   const filteredItems = items.filter(
     (item) =>
@@ -59,4 +61,4 @@ export const ItemList: React.FC<{
       </ul>
     </div>
   );
-};
+});
