+++ src/types/index.ts
@@ -18,12 +18,15 @@ export interface Notification {
 }
 
 export interface AppContextType {
-  theme: string;
-  toggleTheme: () => void;
   user: User | null;
   login: (email: string, password: string) => void;
   logout: () => void;
   notifications: Notification[];
   addNotification: (message: string, type: Notification["type"]) => void;
   removeNotification: (id: number) => void;
+}
+
+export interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
 }
